Add task to publish coronavirus routes
We want a minor update task too, because we don't necessarily want to
send emails if we're just updating the description that appears in search.

At present this doesn't affect anything on the page itself, this is for
publishing the route and appearing in search results (hence the "answer"
document type).