Make special route appear in search
The special_route format does not appear in search, because it may be
used for routes which don't present information to users (like endpoints
for POST requests, or prefix routes which are then handled by apps downstream).

Using the answer format will allow this route to appear in search with a
decent boost. This will also mean that it appears in sitemap.xml which
helps external visibility.

I've used the description from the metatags on the page.