Remove special route for postcode checker
The local restrictions postcode checker hasn't been live since the
begining of January when England went into a national lockdown. At the
moment it looks unlikely that local restrictions will come back again in
their previous form and it's not good for the code base to keep stale
code lying around. The code for the postcode checker is being removed in
https://github.com/alphagov/collections/pull/2380.