Add `/search/latest` route
In https://github.com/alphagov/finder-frontend/pull/2800, a new route
was introduced the replace the "Latest Documents" page that is currently
rendered by Whitehall. This route acts as a redirect to an existing
finder, so will never be published by an application.

This adds the content item, so the finder can be accessed.