Add special routes for past foreign secretaries
As this is a prefix route, it will handle both the individual pages at /government/history/past-foreign-secretaries/:id and the index page.
For now, this keeps the rendering application as whitehall, in order to get a content item live. When we are ready to switch over rendering to collections, we will change the rendering app here and republish.

This also adds parent links to the content item, which will enable us to render breadcrumbs.