Add content item for Past Chancellors page
This is currently rendered by Whitehall, but will be migrated to
Collections.

We need a content item so we can test the rendering in Collections
before migrating the public rendering.

The parent link is being added to ensure the breadcrumb is correctly
displayed on the page.

In a later PR, we will switch the rendering app to Collections.