Add special route for licences API
In https://github.com/alphagov/licence-finder/pull/1275 an API endpoint
was exposed to return the id, location data and the sectors that each
licence in Licence Finder is tagged to.

This data will be used to tag licences to the correct sectors when they
are migrated from Publisher to Specialist Publisher so that they can
be searched for using a specialist finder.

This special route allows the API results to be display. Though the data
being is populated by a database call, the page itself doesn't have a
content item, and needs a way for router to find it.