Add calendars as special routes
These pages were previously published by frontend, but the majority of their information is hard-coded, and not included in the content item. By publishing these as special routes, we will be able to remove all publishing functionality from frontend, removing the architectural oddity of a frontend application also acting as a publishing application, and also retire the calendars schema which is unecessary.