Move homepage into a separate file
The homepage causes a lot of republishing activity through dependency resolution. Separating it from the rest of the routes means that we won't be republishing it every time all routes are republished, therefore decreasing the chance that we accidentally block up the publishing queues with a lot of unecessary activity.